Rename: 회원정보 찾기, 아이디 찾기라는 이름에 맞게 컴포넌트, api 네이밍 수정
