+++ frontend/src/pages/auth/find/components/form/index.ts
@@ -1,3 +1,3 @@
-export { FindForm } from './FindForm'
+export { FindUserIdForm } from './FindUserIdForm'
 export { ResetPasswordForm } from './ResetPasswordForm'
 export { FindType } from './FindType'
