+++ frontend/src/routes/Router.tsx
@@ -14,7 +14,7 @@ import {
   EditActivityPostPage,
   EditBoardPage,
   EditNoticePostPage,
-  FindPage,
+  FindAccountPage,
   LoginPage,
   MainPage,
   MakerPage,
@@ -48,7 +48,7 @@ export const Router = () => {
         <Route path="/auth" element={<AuthRoute />}>
           <Route path="login" element={<LoginPage />} />
           <Route path="signup" element={<SignupPage />} />
-          <Route path="find" element={<FindPage />} />
+          <Route path="find" element={<FindAccountPage />} />
         </Route>
 
         <Route path="/admin" element={<AdminRoute />}>
