+++ frontend/src/service/api/auth/find.ts
@@ -6,7 +6,10 @@ import {
   VerifyResetPasswordCodeRequest,
 } from '@/service/model'
 
-export const findApi = async ({ studentNumber, name }: GetUserIdRequest) => {
+export const findUserIdApi = async ({
+  studentNumber,
+  name,
+}: GetUserIdRequest) => {
   const findClient = new Users(BACKEND_API)
   const response = await findClient.getUserId({ studentNumber, name })
